Naming + types

+++ include/Vessel.hpp
@@ -10,17 +10,16 @@
 namespace stochastic {
 
 class Vessel {
-  std::vector<Reaction> reactions;  // TODO: change data structure
-  SymbolTable<Species, double> species;
+  std::vector<Reaction> reactions;
+  SymbolTable<Species, int> species;
 
  public:
-  const Species add(const std::string& name, double count);
+  const Species add(const std::string& name, int count);
   void add(const Reaction& reaction);
   const Species environment();
   const std::vector<Reaction> getReactions() const;
   const std::vector<Species> getSpecies() const;
-  const SymbolTable<Species, double> getSymbolTable()
-      const;  // TODO: Fix these functions
+  const SymbolTable<Species, int> getSymbolTable() const;
   std::vector<std::string> getSpeciesNames() const;
   double estimateMaxTimesteps(double endTime) const;
 };
